Added better documentation, fixed a few bugs
Updated the documentation to be quite a bit better. I also updated the
directive to support multiple video backgrounds on the same page, added
an end property so that you can optionally end a video early and only
play a short segment, and fixed an issue where the video was not
centering if the width of the container was smaller than the width of
the video.